requirements.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
openai-whisper==20231117
flask==3.0.0
gunicorn==21.2.0
requests==2.31.0

